requirements.txt
BeautifulSoup==3.2.1
Django==1.4
PIL==1.1.7
anyjson==0.3.3
dj-database-url==0.2.1
flickrapi==1.4.2
gunicorn==0.17.4
httplib2==0.8
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.5
python-dateutil==2.1
python-instagram==0.8.0
simplejson==3.1.3
six==1.3.0
wsgiref==0.1.2
